Modify failing test so it passes
In `src/state.rs`, the `login_with_env` command returns a `Result`. I
wanted to make sure it returned an `Err` when no username was set, so I
slapped `unwrap()` on the end and told Rust that the test should panic.
That didn't work. Now I'm `assert`ing that it's an error and the test
passes. Go figure.